requirements.txt
# Core dependencies
sqlalchemy>=1.4.36,<2.0.0  # Compatible with Airflow 2.10.5
psycopg2-binary>=2.9.9  # PostgreSQL adapter
python-dotenv>=1.0.0    # For environment variable management

# Airflow dependencies for local development
apache-airflow==2.10.5  # For local DAG development
apache-airflow-providers-postgres>=5.7.1  # For PostgreSQL connections
apache-airflow-providers-common-sql>=1.7.1  # For SQL operations

# Development dependencies
pytest>=7.4.0
black>=23.7.0          # Code formatting
flake8>=6.1.0         # Code linting
mypy>=1.5.1          # Type checking

# Additional utilities
pandas>=2.0.0         # For data manipulation if needed
numpy>=1.24.0         # For numerical operations if needed 

langchain>=0.1.0        # For LLM integration
langchain-community>=0.0.10  # For Ollama integration

sentence-transformers>=2.2.2  # For BioBERT integration
langchain-huggingface>=0.0.5  # For HuggingFace embeddings
einops>=0.8.1    # Required for Nomic BERT model tensor operations

pyspark==3.5.5  # For Spark integration
matplotlib==3.10.0
oracledb==3.1.0
mlxtend==0.23.4
xgboost==3.0.1
lightgbm==4.6.0
pytorch-tabnet==4.1.0
langchain-ollama==0.3.2
python-Levenshtein==0.27.1
xxhash==3.5.0
hdfs==2.7.3
pingouin==0.5.5
paramiko==3.5.1
boto3==1.38.21
pyarrow==20.0.0
